Extract my PhantomX Joystick changes from PhantomX branch and migrate to the master branch.
Several changes here:
1) The ability to add mappings of the Joysticks button and axes for different
joysticks.  I use PS3 on NUC and PS4 on ODroid XU3 lite.

You can do this by adding parameters to the joy_teleop launch file.

2) The ability to define robot start and stop to same button.  On PS3 I like to
use start button for this.  Again this can be done in launch file.

3) Added support for properly know when push button was pressed versus held
down.  When I had earlier on/off push buttons like the High step state, ran into
issues if you held down button, it would go on and off and on and off...

4) Got rid of main loop and do everything on callback.

5) I wanted to easily add push buttons.  In this case I wanted to add a servo
debug timing state, such that I could sometimes calculate and print out some timing
information (part of a future message...), but I could also see other cases where
I may want to add easy to add buttons, like maybe a button to turn on my IMU
or maybe Lidar Lite scanning.  So I added the ability to easily add buttons,
both in code, where simply add in the button index number to one list and the
ROS topic in another.  Likewise I added the ability to read these in from
launch file.  In case of topics, to either replace current set or to append on
to the list.

Anyway I first did this with PhantomX branch which is working for me, so
extracted the code and converted to master code as to make it easier to pull
if desired.

Here is the part of my launch file on the NUC running off of the main branch
with these changes:
    <node name="Hexapod_Teleop_Joystick" pkg="hexapod_teleop_joystick" type="hexapod_teleop_joystick" />
        <rosparam>
            JOYSTICK_AXES_MAPPING: [0, 1, 2, 3]
            JOYSTICK_BUTTON_MAPPING: [3, 3, 10, 13]
            JOYSTICK_EXTRA_BUTTON_TOPICS_APPEND: ["servo_debug"]
            JOYSTICK_BUTTON_MAPPING_APPEND: [ 12 ]
        </rosparam>
I have it set more or less default for PS3, you can remove axes maping as
it is default.  The BUTTON mapping is not completely standard as I have
the PS3 start button doing start and stop.  default is [3, 0, 10, 13]
Note: 13 is for high stepping stuff which I left in table as figured it did not
hurt anthing if no one is listening for it... I then added my servo_debug
topic (std_msgs::Bool), which uses the button that is button 13, which on
my PS3 is the triangle.

Note: this standard button stuff could be functionally enhanced more. Currently
I default that all of these buttons start off with the value of false.
Could add support to read in initial values as well...   Also could add support
to instead of send toggle like messages, could send message when button is pressed
and another when released... So far I had no need for this.